changed maany things, check this one
Fixed some minor bugs:
cars went too far out of bound
sometimes if many fitnesses were negative, they would get = 0, changed it.

Made a new generation after a certain number of steps, so I dont need to press space that much (it is just for testing)

Plots the Fitness vs generation after the program finishes. 
(Here be careful because I made it so that the duration of each generation increases with time ...(count == 300 + generation_n * 20)... )
So fitness may be increasing just because the duration is increasing.
You can ofc change it

Pygame screen also shows generation number

Now each generation the selected parents are kept identical and they have some children too.
So some from the old generation remain, instead of all new babies 
I think the prof wanted us to do that, either way we can test which option is better

Also added a function to test weather each car is crashing with another and how to include it in the fitness, I know the prof would like that.
let's leave it for later tho.

The biggest problem for me is that cars tend to go in circles too much. 
Even if I penalize being out of the road WAY more than the benefit of distance traveled.